refactor: 테스트 페이지 route 구조 변경
